cb-21-deployment/requirements.txt
easyocr==1.6.2
Flask==2.2.2
nltk==3.6.5
openai==0.26.0
pdf2image==1.16.0
Pillow==9.4.0
torch==1.13.0
transformers==4.25.1
Werkzeug==2.2.2
gunicorn
